Fixed dwca-codegen namespace for Terms.Attributes Added README.md to dwca-codegen Re-organized mappers
